ai_api/requirements.txt
langchain-core
langchain-ollama
langchain
chromadb
python-dotenv